Add delete functionality to boards Write feature test for deleting a board
