+++ .rspec
@@ -1,2 +1,3 @@
 --color
+--format=documentation
 --require spec_helper
